LayoutTests/http/tests/inspector/console-resource-errors.html

<html>
<head>
<script src="console-test.js"></script>
<script src="inspector-test.js"></script>

<link rel="stylesheet" type="text/css" href="missing.css">

<script>
var xhr = new XMLHttpRequest();
xhr.open("GET","non-existent-xhr", false);
xhr.send(null);

function performDownload()
{
    var a = document.body.getElementsByTagName("a")[0];
    var event = document.createEvent("UIEvents");
    event.initUIEvent("click", true, true, window, 1);
    a.dispatchEvent(event);
}

function test()
{
    function onResource(event)
    {
        var resource = event.data;
        if (/\/empty\.zip$/.test(resource.url)) {
            InspectorTest.expandConsoleMessages();
            InspectorTest.dumpConsoleMessagesWithClasses(true);
            InspectorTest.completeTest();
        }
    }
    function onMainResourceCommitLoad()
    {
        WebInspector.console.clearMessages();
        WebInspector.networkManager.addEventListener(WebInspector.NetworkManager.EventTypes.ResourceFinished, onResource, this);
    }
    WebInspector.networkManager.addEventListener(WebInspector.NetworkManager.EventTypes.MainResourceCommitLoad, onMainResourceCommitLoad, this);
    InspectorTest.reloadPage(function() {
        InspectorTest.evaluateInPageWithTimeout("performDownload();");
    });
}

</script>
</head>

<body onload="runTest()">
<p>
<script src="non-existent-script.js"></script>

<iframe src="non-existent-iframe.html"></iframe>

Tests that errors to load a resource cause error messages to be logged to console.
</p>

<iframe name="frame"></iframe>
<a href="resources/empty.zip" target="frame">download</a>
</body>
</html>
