LayoutTests/http/tests/inspector/console-resource-errors-expected.txt

 Tests that errors to load a resource cause error messages to be logged to console.

 download
Page reloaded.
missing.cssGET http://127.0.0.1:8000/inspector/missing.css 404 (Not Found) console-message console-error-level
non-existent-iframe.htmlGET http://127.0.0.1:8000/inspector/non-existent-iframe.html 404 (Not Found) console-message console-error-level
non-existent-script.jsGET http://127.0.0.1:8000/inspector/non-existent-script.js 404 (Not Found) console-message console-error-level
non-existent-xhrGET http://127.0.0.1:8000/inspector/non-existent-xhr 404 (Not Found) console-message console-error-level
